server/requirements.txt
fastapi==0.115.8
fastapi-cli==0.0.7
uvicorn==0.34.0
mysql-connector-python==8.0.31
pyodbc>=4.0.35
numpy==2.0.0
pandas>=1.5.3
pyjwt==2.10.1
cryptography==44.0.2
huey==2.5.3
redis==5.*
pydantic-settings==2.7.1
zarr>=2.18.0
python-multipart==0.0.20
passlib==1.7.4
argon2_cffi==25.*
PyJWT>=2.8.0
